Safari-609.1.11.1/LayoutTests/fast/dom/Window/child-window-focus-expected.txt

Test that a Window can be focused / closed by its opener

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Opening new window
PASS newWindow.opener is window
newWindow.focus()
PASS New Window focused
PASS newWindow.closed is true
PASS successfullyParsed is true

TEST COMPLETE
